added persistent terminal utility + other changes
Added a function in utils that allows terminal to stay open when the
command passed to the -e option exits. In edit.sh and experiment.sh, the
terminal and shell now stays when the nvim process quits.

Added short docs for some functions in utils

Changed update.sh to a function in aliases