rmpc: restored keymap to play a single song
An update changed the Enter key to also add the entire playlist into
queue in addition to playing the selected song. I added Shift + Enter
which does the old Enter function, playing only the selected song,
without adding the entire playlist into queue.

`l` also used to add only the selected song to queue, but now it adds
the entire playlist. But the `a` keymap already does the old `l`
function, so no need to add a new keymap.